Final fix of updater

+++ ScribeBot/Interface/Window.cs
@@ -308,9 +308,10 @@ namespace ScribeBot.Interface
 
         private void downloadUpdate_Click(object sender, EventArgs e)
         {
-            File.Copy("Updater.exe", $"Updater {Core.Version}.exe");
+            if(!File.Exists($"Updater {Core.Version}.exe") && File.Exists("Updater.exe"))
+                File.Copy("Updater.exe", $"Updater {Core.Version}.exe");
 
-            Process.Start("Updater {Core.Version}.exe");
+            Process.Start($"Updater {Core.Version}.exe");
 
             Core.Close();
         }
